Bugfix: issue #613 (ClassCastException during read) in gradle-restructure branch
git-svn-id: https://jmonkeyengine.googlecode.com/svn/branches/gradle-restructure@11077 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ jme3-core/src/main/java/com/jme3/shadow/AbstractShadowRenderer.java
@@ -610,7 +610,7 @@ public abstract class AbstractShadowRenderer implements SceneProcessor, Savable 
         InputCapsule ic = (InputCapsule) im.getCapsule(this);
         assetManager = im.getAssetManager();
         nbShadowMaps = ic.readInt("nbShadowMaps", 1);
-        shadowMapSize = ic.readInt("shadowMapSize", 0);
+        shadowMapSize = ic.readFloat("shadowMapSize", 0f);
         shadowIntensity = ic.readFloat("shadowIntensity", 0.7f);
         edgeFilteringMode = ic.readEnum("edgeFilteringMode", EdgeFilteringMode.class, EdgeFilteringMode.Bilinear);
         shadowCompareMode = ic.readEnum("shadowCompareMode", CompareMode.class, CompareMode.Hardware);
